+++ common/src/client/java/io/github/jamalam360/jamlib/client/impl/mixin/event/ClientLevelMixin.java
@@ -8,7 +8,7 @@ import org.spongepowered.asm.mixin.injection.Inject;
 import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
 
 @Mixin(ClientLevel.class)
-public class ClientLevelMixin {
+public abstract class ClientLevelMixin {
 	@Inject(
 			method = "tickEntities",
 			at = @At("HEAD")