+++ common/src/client/java/io/github/jamalam360/jamlib/client/impl/mixin/event/MouseHandlerMixin.java
@@ -2,8 +2,11 @@ package io.github.jamalam360.jamlib.client.impl.mixin.event;
 
 import com.llamalad7.mixinextras.sugar.Local;
 import io.github.jamalam360.jamlib.client.api.events.ClientMouseScrollEvents;
+import io.github.jamalam360.jamlib.client.impl.keymapping.ScreenKeyMappingHelper;
 import net.minecraft.client.Minecraft;
 import net.minecraft.client.MouseHandler;
+import net.minecraft.client.input.MouseButtonEvent;
+import net.minecraft.client.input.MouseButtonInfo;
 import org.spongepowered.asm.mixin.Final;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.Shadow;
@@ -12,7 +15,7 @@ import org.spongepowered.asm.mixin.injection.Inject;
 import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
 
 @Mixin(MouseHandler.class)
-public class MouseHandlerMixin {
+public abstract class MouseHandlerMixin {
 	@Shadow
 	@Final
 	private Minecraft minecraft;
@@ -44,4 +47,18 @@ public class MouseHandlerMixin {
 			ci.cancel();
 	    }
     }
+
+	@Inject(
+			method = "onButton",
+			at = @At(
+					value = "INVOKE",
+					target = "Lnet/minecraft/client/gui/screens/Screen;afterMouseAction()V"
+			),
+			cancellable = true
+	)
+	private void jamlib$triggerScreenKeybinds(long handle, MouseButtonInfo rawButtonInfo, int action, CallbackInfo ci, @Local MouseButtonEvent ev) {
+		if (ScreenKeyMappingHelper.onMousePressed(ev)) {
+			ci.cancel();
+		}
+	}
 }