+++ CHANGELOG.md
@@ -1,2 +1,3 @@
 - Add a resource reload listener registry.
 - Add container render events.
+- Add an option for registered keybinds to also trigger while in screens.